requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
attrs==21.2.0
    # via jsonschema
colorama==0.4.4
    # via rich
commonmark==0.9.1
    # via rich
jsonschema==3.2.0
    # via -r requirements.in
pygments==2.9.0
    # via rich
pyrsistent==0.17.3
    # via jsonschema
pyyaml==5.4.1
    # via -r requirements.in
rich==10.3.0
    # via -r requirements.in
six==1.16.0
    # via jsonschema

# The following packages are considered to be unsafe in a requirements file:
# setuptools
